Move XTS from libdecrepit to libcrypto.
* Removed OPENSSL_EXPORT from capability function declaration
  in AArch64
* Remove the no-longer-needed capability function in x86.
