+++ crypto/CMakeLists.txt
@@ -452,6 +452,7 @@ add_library(
   x509v3/v3_purp.c
   x509v3/v3_skey.c
   x509v3/v3_utl.c
+  xts/xts.c
 
   ${CRYPTO_ARCH_SOURCES}
 )
@@ -635,6 +636,7 @@ if(BUILD_TESTING)
     x509/x509_test.cc
     x509/x509_time_test.cc
     x509v3/tab_test.cc
+    xts/xts_test.cc
 
     $<TARGET_OBJECTS:crypto_test_data>
     $<TARGET_OBJECTS:boringssl_gtest_main>
