+++ crypto/xts/xts.c
@@ -54,8 +54,8 @@
 #include <openssl/cipher.h>
 #include <openssl/err.h>
 
-#include "../crypto/fipsmodule/modes/internal.h"
-#include "../crypto/fipsmodule/aes/internal.h"
+#include "../fipsmodule/modes/internal.h"
+#include "../fipsmodule/aes/internal.h"
 
 typedef struct xts128_context {
   AES_KEY *key1, *key2;
